// src/app/models/reference.ts
export class RefModel {
    id?: number;
    code?: string;
    name?: string;
    subId?: number;
    subName?: string;
    status?: string;
    amount?: any;
    sortOrder?: number;
}
export class PujaModel {
    id?: number;
    amount?: any;
    outsideAmount?: any;
    pujaHours?: number;
    pujaName?: string;
    pujaDetails?: string;
    status?: string;
    sortOrder?: number;
}